backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
playwright
asyncio
aiohttp
openai
beautifulsoup4
pytest
gunicorn
modal
